+++ module/ldbc-connector/shared/src/main/scala/ldbc/connector/net/protocol/CallableStatement.scala
@@ -18,7 +18,7 @@ import cats.effect.*
 import org.typelevel.otel4s.Attribute
 import org.typelevel.otel4s.trace.{ Tracer, Span }
 
-import ldbc.sql.{ ResultSet, DatabaseMetaData }
+import ldbc.sql.{ ResultSet, DatabaseMetaData, ParameterMetaData }
 
 import ldbc.connector.*
 import ldbc.connector.data.*
